改善batch render，提高支持绘制的精灵数量

+++ Src/apengine/graphics/buffers/indexbuffer.cpp
@@ -1,6 +1,6 @@
 #include "indexbuffer.h"
 
-IndexBuffer::IndexBuffer(GLushort* data, GLsizei count)
+IndexBuffer::IndexBuffer(GLuint* data, GLsizei count)
 	:m_Count(count),m_Data(data)
 {
 	m_BufferID = 0;
@@ -29,7 +29,7 @@ void IndexBuffer::init()
 {
 	glGenBuffers(1, &m_BufferID);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_BufferID);
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, m_Count * sizeof(GLushort), m_Data, GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, m_Count * sizeof(GLuint), m_Data, GL_STATIC_DRAW);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
 }
